Move document.py to repo layer
